feat: Add Client and Supplier types and mock data
Introduces new data types for clients and suppliers to the EBF Manager application. This change also includes mock data for these new types, which will be used for initial testing and development of related features. The `Transaction` type in `types.ts` has also been updated to include the `site` property.